Add email to donations

+++ db/migrate/20121129020939_drop_then_create_follows.rb
@@ -1,6 +1,6 @@
 class DropThenCreateFollows < ActiveRecord::Migration
   def up
-  	drop_table :follows
+  	drop_table :follows if table_exists?(:follows)
   	create_table :follows do |t|
       t.integer :volunteer_id
       t.integer :organization_id
@@ -10,5 +10,9 @@ class DropThenCreateFollows < ActiveRecord::Migration
   end
 
   def down
+  end
+
+  def self.table_exists?(name)
+    ActiveRecord::Base.connection.tables.include?(name)
   end
 end
